Allow embedding local images

+++ pypub/chapter.py
@@ -60,6 +60,12 @@ def save_image(image_url, image_directory, image_name):
     if image_type is None:
         raise ImageErrorException(image_url)
     full_image_file_name = os.path.join(image_directory, image_name + '.' + image_type)
+
+    # If the image is present on the local filesystem just copy it
+    if os.path.exists(image_url):
+        shutil.copy(image_url, full_image_file_name)
+        return image_type
+
     try:
         # urllib.urlretrieve(image_url, full_image_file_name)
         with open(full_image_file_name, 'wb') as f:
